Phase 1: clear pointer-width and signal-handler warnings
Knock down 38 of the 92 warnings remaining after the compile-success
milestone. The remaining 54 are all in alloc.c (31, Phase 1c struct
sizing), eval2.c (10, Phase 1g callg ABI), divbig.c/pbignum.c (10,
Phase 1h bignum redesign), and global.h (2 cascading from divbig).

Per-file changes:

  * fex3.c:  Ndumplisp gated under linux_x86_64 -- dumplisp is
    deferred (PortPlan.md), and the original implementation pokes
    a.out internals. Stub returns nil. Common aout.h include moved
    to top so Lgetaddress (still compiled) sees struct nlist.
  * fex4.c:  ctim widened to time_t (matches time/ctime/localtime
    on glibc); local libc redeclarations of ctime/localtime removed
    (they live in <time.h>). Lsyscall args[] widened to long. badmr
    forward decl uses SIGTYPE.
  * ffasl.c: verify() lifted out of the linux_x86_64 gate -- it's
    used all over the kernel and being in ffasl.c was incidental.
  * inits.c: badmr returns SIGTYPE; re_enable replaces deprecated
    sigblock/sigsetmask with sigprocmask on Linux. (uintptr_t)
    cast on sbrk(0) result.
  * lam6.c:  sginth fwd-decl uses SIGTYPE.
  * lam7.c:  <time.h> included for ctime/time. (char *)0 -> NULL.
    Lgetfb's (int)temp routed through (intptr_t) (still truncates,
    flagged as Phase 1c TODO).
  * lam8.c:  vector size word read via (int)(intptr_t)v->...;
    sprintf-via-cast block flagged as Phase 1g TODO with diagnostic
    pragma (one residual warning the pragma can't reach is left).
  * lam9.c:  verify() casts to (char *) at chmod/stat/getpwnam call
    sites since verify returns a Lisp string-bearing pointer.
  * lamgc.c: GC trace-output address truncations cast through
    uintptr_t; the gc.out file isn't read by anything in this build
    so a 32-bit lowdata field is acceptable.
  * lamr.c:  Lmaknum's (int)ptr -> (int)(intptr_t)ptr (Phase 1c
    TODO, still truncates). lctrace declared lispval (was int) in
    data.c so Llctrace round-trip stops truncating.
  * fpipe.c: comment-after-#endif fixed.

+++ franz/lamgc.c
@@ -61,7 +61,13 @@ Lgcstat()
 	    case 0:  if((gcstat = creat("gc.out",0644)) < 0)
 		       error("cant open gc.out",FALSE);
 		     hhh.version = 5;
-		     hhh.lowdata = (int)beginsweep;
+		     /* hhh.lowdata is an int field; the address may not
+		      * fit on 64-bit. The gc.out file is not consumed by
+		      * anything in this build, so a truncated value is
+		      * acceptable -- but we cast through uintptr_t first
+		      * to silence the warning explicitly.
+		      */
+		     hhh.lowdata = (int)(uintptr_t)beginsweep;
 		     printf("writing %d bytes \n",sizeof(hhh));
 		     write(gcstat,(char *)&hhh,sizeof(hhh));
 		     gccount = 0;
@@ -75,7 +81,7 @@ Lgcstat()
 				  sizeof(nbytes));
 		     write(gcstat,(char *)&nbytes,sizeof(nbytes));
 		     write(gcstat,(char *)&typetable[ATOX(beginsweep)+1],
-				nbytes = ((int)datalim - (int)beginsweep)>>9);
+				nbytes = ((uintptr_t)datalim - (uintptr_t)beginsweep)>>9);
 		     printf("writing %d bytes \n",nbytes+sizeof(nbytes));
 		     write(gcstat,(char *)&nbytes,sizeof(nbytes));
 		     close(gcstat);
